+++ src/main/java/ru/mystamps/web/service/SeriesServiceImpl.java
@@ -154,7 +154,7 @@ public class SeriesServiceImpl implements SeriesService {
 	
 	@Override
 	@Transactional(readOnly = true)
-	public int countByMichelNumber(String michelNumberCode) {
+	public long countByMichelNumber(String michelNumberCode) {
 		Validate.isTrue(michelNumberCode != null, "Michel number code must be non null");
 		Validate.isTrue(!michelNumberCode.trim().isEmpty(), "Michel number code must be non empty");
 		
@@ -163,7 +163,7 @@ public class SeriesServiceImpl implements SeriesService {
 	
 	@Override
 	@Transactional(readOnly = true)
-	public int countByScottNumber(String scottNumberCode) {
+	public long countByScottNumber(String scottNumberCode) {
 		Validate.isTrue(scottNumberCode != null, "Scott number code must be non null");
 		Validate.isTrue(!scottNumberCode.trim().isEmpty(), "Scott number code must be non empty");
 		
@@ -172,7 +172,7 @@ public class SeriesServiceImpl implements SeriesService {
 	
 	@Override
 	@Transactional(readOnly = true)
-	public int countByYvertNumber(String yvertNumberCode) {
+	public long countByYvertNumber(String yvertNumberCode) {
 		Validate.isTrue(yvertNumberCode != null, "Yvert number code must be non null");
 		Validate.isTrue(!yvertNumberCode.trim().isEmpty(), "Yvert number code must be non empty");
 		
@@ -181,7 +181,7 @@ public class SeriesServiceImpl implements SeriesService {
 	
 	@Override
 	@Transactional(readOnly = true)
-	public int countByGibbonsNumber(String gibbonsNumberCode) {
+	public long countByGibbonsNumber(String gibbonsNumberCode) {
 		Validate.isTrue(gibbonsNumberCode != null, "Gibbons number code must be non null");
 		Validate.isTrue(
 			!gibbonsNumberCode.trim().isEmpty(),
